+++ NEWS.md
@@ -48,6 +48,12 @@
 * Fixed issue in `test_predictions()` with wrong order of term labels when
   a focal term was a character vector.
 
+* Fixed issue in `ggpedict()` with `wbm` models from package *panelr*.
+
+* Fixed issue in `ggemmeans()` for `glmmTMB` models with zero-inflation, when
+  `terms` included variables that were specified in the conditional, but not
+  in the zero-inflation model formula.
+
 # ggeffects 1.5.2
 
 ## General